requirements.txt
rich
click
networkx
psycopg2-binary
pandas
tqdm
pyspark
bertopic
transformers
ipython
praw
scikit-learn
sentence-transformers
SQLAlchemy
python-dotenv
nest-asyncio
pyppeteer
requests-html
numpy
